add a line regarding the first run of sgadmin.sh
the -rl flag appears to cause an initial run of sgadmin.sh to fail against new clusters as it expects a non-existent searchguard index. The linked issue discussion mentions the effect regarding the three flags mentioned in this document, however, the sgadmin tool's java code suggests that -rl also may have this effect but it is not reflected in the issue discussion nor the documentation. 

https://github.com/floragunncom/search-guard/issues/228

https://github.com/floragunncom/search-guard/blob/master/src/main/java/com/floragunn/searchguard/tools/SearchGuardAdmin.java#L113